fix bug

+++ backend/src/main/kotlin/com/nour/service/ExamService.kt
@@ -94,7 +94,7 @@ class ExamService(
         ))
         return ExamResultResponse(
             id = UUID.randomUUID().toString(),
-            score = score,
+            score = score.toDouble(),
             isPassed = isPassed,
             correctCount = correct,
             totalQuestions = questions.size,
